Refactor: Modularized Dayplanner into form, landmarks, footer and save function
